src/requirements.txt
pymysql==1.0.2
google-api-python-client==2.0.2
google-auth==1.30.0